resolver: rename response_cache module to cache
